// src/api/generated/api.ts
export interface ArticleResp {
    'id': number;
    'title': string;
    'description'?: string;
    'local'?: string;
    'published_at'?: string;
    'revelance'?: number;
    'thumbnail_id'?: number;
    'thumbnail'?: string;
    'links'?: Array<LinkResp>;
    'attachments'?: Array<AttachmentResp>;
    'skills'?: Array<SkillResp>;
}
export interface AttachmentResp {
    'id': number;
    'title': string;
    'link': string;
    'description'?: string;
}
// ...
export interface LinkResp {
    'id': number;
    'title': string;
    'link': string;
    'description'?: string;
}
// ...
export interface SkillResp {
    'id': number;
    'title': string;
    'description'?: string;
}